refactor(flow): move state management here
